Add selector assertions
This adds two PHPUnit assertion methods and the respective Pest expectations to the two relevant classes `TestView` and `TestResponse`. An instance of `TestView` ist returned e.g. when we render an isolated Blade component, e.g. like this:

```
$view = $this->blade('<x-thumb result="adopted" />');
```

An instance of `TestResponse` is returned when we send a response in our HTTP feature test, e.g. like this:

```
$response = $this->get('/votes/1');
```

In both cases, it would be handy to be able to assert that a particular elements exists using CSS selectors, e.g. like this:

```
$response->assertSelector(‘.thumb—adopted.thumb—circle’)
```

Right now, we work around this using `assertSee`, e.g. like this

```
$response->assertSee(‘thumb--adopted thumb--circle’)
```

This works, but is prone to errors. It simply asserts that the string exists somewhere in the response. As a consequence, it will break if for example the order of classes in the `class` attribute changes, the above `assertSee` matches the first element, but not the second:

```
<span class=“thumb thumb--adopted thumb--circle”>
<span class=“thumb thumb--circle thumb--adopted”>
```

Also, using `assertSee` does not allow for more complex assertions, e.g. to test for nested elements:

```
<li class=“list-item”>
  <span class=“thumb thumb--adopted”></span>
</li>
```

This is now possible using the newly added assertions:

```
$response->assertSelector(‘.list-item > .thumb.thumb—adopted’)
```

Apart from that, it’s also possible to assert that a selector matches an exact amount of elements and that a selector matches elements with a given text:

```
// Expect 3 elements with the class `.thumb--adopted`
$response->assertSelector(‘.thumb--adopted’, 3);

// Expect 2 elements with class `.list-item__subtitle` contain text `Greens/EFA`
$response->assertSelectorText(‘.list-item__subtitle’, ‘Greens/EFA’, 2);
```

As our test framework Pest basically is only a wrapper around PHPUnit, providing a more intuitive interface to write tests, I implmented these as “standard” assertions first, then added the respective Pest expectations – i.e. we can use both in our tests like this:

```
expect($response)->toHaveSelector(‘.thumb--adopted’);
expect($response)->toHaveSelectorWithText(‘.list-item’, ‘Greens/EFA’);
```

The assertions themselves are implemented using the `css-selector-converter` package. PHP has native DOM libraries to parse XML/HTML documents using xPath expressions. The basic idea is to first convert the CSS selector to an xPath expression using `css-selector-converter`, then use the native PHP DOM libraries to query the document using the resulting xPath selector.

Closes 280.