Remove config:cache from composer test script
This was intended as a workaround for the problem that we need to run artisan config:cache after running tests in order to access the local database afterwards. However, this is not a suitable workaround since there is no option to provide arguments to a specific command in a script list as used here. This leads to the name of the test file to execute when running only one test file to the cache:clear command that errors.