Use final votes instead of primary for selection and add final scope
Sometimes the last vote in parliament on a topic is on an amendment. Nevertheless, these votes are subsequently the final position of the parliament on the given topic. Thus, relying on the primary category to identify the main votes is not sufficient. We instead use the last vote that was held on each subject.