Without search query, sort voting lists by date
If a user didn’t yet enter a search term, we want to display a list of votes in inverse chronological order, grouped by plenary session.

MeiliSearch only supports custom ranking rules for numeric attributes, that’s why we need to index the date as a Unix timestamp (as opposed to the date-time string we indexed before).