Fix automated scraper runs
Our current logic to automatically import new vote results doesn't work out… 🙃 Each Sunday, we’re trying to import vote results for the current week (i.e. the previous seven days). However, all documents are usually only available in the week following the plenary session.

This commit changes the `scrape:all` command.

It doesn't accept `from` and `to` options anymore. Instead, it tries to find the oldest session that has no associated votes and uses that session’s start and end date to scrape votes and voting lists.

That way, no matter when both votes and voting lists become available, we’ll import them. There’s one problematic case: If there’s a session without votes (be that a valid session or a session that for some reason didn’t actually take place), we’d try to scrape votes for that session again and again… Even if there are more recent sessions.

I’d simply accept that for now and see if that actually is a problem.