Add switch to manually exempt Sessions from voting list scraping
This fixes a crash that occurs when we try to scrape the voting lists for a Session that has no voting lists in scrape:all. In the future, we can manually mark such Sessions. I opted for manual marking as the implications of exempting a Session are quite severe and this case is very rare.