Upgrade to php8.1
Installing `composer` from the `alpine` repositories lead to a second `php` version being installed which was always used by `composer` as default. Due to a delay in packaging, this lead to `composer` still using `php8.0`.
This approach to installing `composer` is taken from https://getcomposer.org/doc/00-intro.md#docker-image.

`phpunit` needed to be updated due to a compatibility issue with `php8.1`, which made upgrades to `pest` necessary as well.