Reflect change on Member page of EP in name we return in scrapers
In the past, the EP Page returned Miranda Paz as her name, which we needed to reflect in our VotingLists, as we match them based on name. The name on the EP page was changed and is now consistent with the name as it is found in the VotingLists.
We switched to the new name automatically when scraping the changes in member profiles, and this change is now necessary to successfully scrape VotingLists again.