Upgrade to Meilisearch 0.28.1
This allows us to index nested objects. I’d like to reuse our current search UI on member profiles as well. This has two advantages:

1) The votes displayed on member profiles are often out of date, because we cache them statically. When dynamically rendering votes, the most recent votes will always be “fresh”, as they are fetched from the search index.

2) We can search for a certain keyword or topic and see at a glance how an MEP voted in multiple votes matching the search term.

With nested index objects, we can index a vote like this (where 1234 and 5678 are the IDs of members):

```
{
  "title": "Lorem Ipsum",
  "result": "ADOPTED",
  …
  "members": {
    "1234": "FOR",
    "5678": "AGAINST",
    …
  }
}
```

We can then retrieve only the voting for a single member from the search index:

```
{
  "q": "Lorem",
  ...
  "attributesToRetrieve": [
    "title",
    "members.1234",
    ...
  ]
}
```