Display votes in MEP profile only if MEP was an active member
I would have preferred to implement this using a filter like this:

```
members.123 exists
```

or

```
members.123 = 'FOR' or members.123 = 'AGAINST' or members.123 = 'ABSTENTION' or members.123 = 'NOVOTE'
```

However, filtering requires us to specify a list of attributes in the index settings (and that index setting doesn't support wildcards). As the attributes that need to be filterable, this isn't possible.

Instead, I changed the way we index voting lists:

* `members`: an array of member IDs that have voted (or could have voted)
* `votings`: an object that maps member IDs to vote positions

In order to fetch voting lists for a specific MEP, we can now simply add a filter like this:

```
members = 123
```