FIx `ParliamentSessionsScraper` for Sessions that span multiple months
We previously assumed, that a session never spans multiple years or months. This assumption does not hold, and the first such session that we  encountered lead to a faulty session entry that in turn got displayed on the landing-page in a nonsensical way.
This removes the assumption of a session never spanning multiple months. For years, we still keep our assumption