Add basic CSV export
We used to simply publish dumps of our database for anyone wanting to run analysis on the entire dataset. However, we store some data in denormalized form. This can make it much more difficult to work with the dataset.

This implements a simple CSV export for members, groups, countries, votes as well as relationships between these record types. The individual CSV tables are mostly normalized so they are easy to work with with tools such as Pandas, R, etc.

In addition to the CSV tables, exports also include metadata about each table following the CSV on the Web standard[0].

[0]: https://www.w3.org/TR/tabular-data-primer/