Refactor worker
Previously, the worker executed only pipelines. We wrap pipeline handlers with additional error handling code and log every pipeline run in the database. This is necessary, because in some cases, executing a pipeline may depend on the result of a previous pipeline run.

The worker is now also executing non-pipeline jobs, for example generating CSV exports. We do not need to log these in the database etc., so I’ve slightly refactored the worker: The `schedule` method schedules a job, and the `schedule_pipeline` method essentially is a thin wrapper around the `schedule` method that also applies the additional error handling and logging for pipelines.