Migrate to fetch client package
Previous versions of `openapi-ts` would generate TypeScript types based on the API schema *and* client code to send requests. Recent versions of `openapi-ts` only generate the TypeScript types and a few helper methods to send API requests, but not the actual code to send the API requests. Instead, this is provided as a separate package (e.g. `@hey-api/client-fetch`).