Configure `noLabelWithoutControl` rule
This is a new rule that is enabled by default. It checks that every `label` element is associated with a form input, either by using the `for` attribute or by nested the form input inside of the label.

In our case we nest the form inputs inside of the label. However, Biome detects only `input` etc. elements rendered directly, and we use a custom component that in turn renders the actual form input.

Thankfully, Biome can be configured to handle this use case: https://biomejs.dev/linter/rules/no-label-without-control/#options