Downgrade Node.js types
This was causing a TypeScript error in the `@tinyhttp/res` dependency:

```
Cannot find type definition file for 'node/http.js'
```

I’m not entirely sure what’s happening here, so the quick fix is to revert this for now. I think this would also likely be resolved by updating tinyhttp to the latest version.