Fix group memberships test so that it can run against live data source
This test was using an outdated fixture. As the 9th plenary term is over, there are no ongoing group memberships anymore and all group memberships now have an end date. I've updated the test to use Markus Weber's group memberships from the current term. We will have to update the test when something changes (or for the 11th term the latest), but that should be manageable.