Update main vote analysis to also handle special vote titles
Until now, we’ve decided whether a vote is a main vote solely based on the description of a vote. However, in rare cases a main vote may not have description and only a title (e.g. in case of the election of the Commission). I’ve updated the analyzer to handle votes based on title in addition to the existing heuristic.