Fix the EuroVoc ID
We previously used the DC Terms `identifier` property. However, this property is sometimes ambiguous. For example, the domain concept "AGRICULTURE, FORESTRY AND FISHERIES" has the DC Terms identifier `56`, but the correct ID is `100156`. Instead, we now extract the ID from the resource URI (`http://eurovoc.europa.eu/100156` in this example).