Simplify per-field boost factors
Previously, `rapporteur` was the only field that had the default boost factor of 1. Now, `geo_areas`, `eurovoc_concepts`, and `rapporteur` all have the same default factor of 1. In practice, this shouldn’t make a difference, as the names of rapporteurs are unlikely to appear in any of the other fields anyway.