Fix types
I’m not entirely sure why this started failing only now and not already when we merged the Xapian changes. The error looks right though, i.e. the function was indeed missing type hints. In this particular case, the solution was to simply remove the types for `Weight.__init__` because that is never called directly.

Fixes #1101.