Extract vote results mentioned in press release texts
We can currently only match press releases with votes if the press release references a report or procedure. For non-legislative procedures, this is usually not the case. However, even press releases about non-legislative procedures usually mention the vote results (how many MEPs voted for/against/abstained). If there is exactly one vote with the same result on the day of the press release, we can use that info to match the press release to the vote.

As a first step, this commit implements changes to scrape the full text of press releases and to extract vote results from the text.