Reenable package mode
With package mode disabled, Poetry won’t install scripts, so the `htv` CLI doesn’t work.

Since Poetry 2, the package name in `pyproject.toml` has to match the folder name. In our case, in the Docker container, we mount the backend files into `/howtheyvote/backend`, so we also have to specify the path to the package directory explicitly.