Always build Docker images
I’ve messed with our Docker setup a few times now because I haven’t tested the changes carefully enough. We currently only build Docker images in CI on demand, but not for every commit/pull request because it seemed unnecessary.

This commit changes that to ensure we catch Docker issues in CI. The image is still only pushed to the registry when running the workflow manually.