Extract logic to calculate number of votes per position
This was previously only used in the votes API endpoint, but will now also be used to match votes to press releases.

I also simplified the logic a bit. There are now two separate functions (instead of just one in the old implementation): The first only computes the counts for a given list of member votes. And the second does the grouping (e.g. based on political group).