Remove `is_featured` column
The meaning of `is_featured` was always a bit confusing for end users, and in the end we only ever used it internally for search results ranking. With #1108, we started storing a reference to a press release in the `votes` table (if there is a press release mentioning a vote). This makes the `is_featured` column redundant, because `is_featured == True` if there is a press release for a vote.

This removes the column everywhere (including from the API and exports). Whether a vote has a corresponding press release is still a ranking factor, but it’s now named `has_press_release` in the search index which should be more obvious.