Fix type checking errors
The issue in these cases is that mypy treats

```py
country = Country[country_code]
```

as a type expression, i.e. as a specialization of a generic class `Country` (a specialization of the generic class `Country`) and not as a value expression (indexing into `Country` class). This will cause mypy to fail because `Country` isn’t generic.

This adds a workaround to trick mypy into treating it as a value expression.

We’re making use of the `Country[country_code]` pattern because it’s similar to how you can access enum memberes (which we were using before), but maybe that’s trying to be a bit too smart and we should just replace it with something simple when we run into this issue again.