Remove phrase subquery
The idea behind this was to boost results that contain the search query in the same order. However, this was causing errors for nested search queries (and queries that were parsed as such), because Xapian supports phrase queries only if all subqueries are leaf queries.

Not sure if there is a good alternative. For now, I’m simply removing this. I’ve also removed the option to use `+` and `-` operators, as these are also commonly used as normal characters in search queries (e.g. date ranges).