Check language of VOT document
If an English translation isn’t yet available, requesting the English translation will return the French original. This checks whether the returned document actually is the English version. If not, it raises a `NoWorkingUrlError` exception, which will be treated in the same way as if the EP website had returned a 404 (i.e. the scraper will be retried later).

Fixes #1167