Add optional idempotency key to scheduled pipelines
If such a key is given, a pipeline will only be executed until it has been executed successfully for the first time.

This allows us to retry a pipeline multiple times per day until a data source becomes available (as we already do for RCV lists).

But it can also be used to retry a pipeline for multiple days (as is needed for VOT lists which are usually published with some delay).