Don’t run test against live data (to fix scheduled scraper tests)
This test ensures that the `VOTListScraper` doesn’t use a French version of the VOT list when the English translation isn’t yet available. There’s no way to test this against live data (except in the short time frame between a session and the time when the translations are published).