Remove data issues
This was originally intended to perform continuous data quality checks. However, in practice, we don't really use this and rely a lot on manual checks. At the same time, this adds quite a bit of complexity to the code base, and adding new checks is more work than it should be.

I think a better way of running these checks would be in a separate system, for example in a GitHub Action based on our data exports.