Remove `VoteGroup` model
There can be multiple votes on the same topic on the same day (for example due to amendments). We refer to all groups that are related in such a way as a single vote group.

Currently, we identify vote groups by computing a hash over the vote date and title. We store this hash as the `group_key` attribute of the `Vote` record, but we also create a `VoteGroup` record.

While the `group_key` on the `Vote` record is useful (for example to fetch amendments for a given vote), we don’t really use the separate `VoteGroup` record at the moment.

In order to remove complexity and unnecessary code, I suggest removing the `VoteGroup` model entirely.