Add support for `contains` operator to `ListType`
In many cases, we use JSON columns to store lists of values, for example `geo_areas` related to a vote. This is implemented using a SQLAlchemy type decorator called `ListType`. This adds support for the `contains` operator to this type.

For example, the following query

```
select(Vote).where(Vote.geo_areas.contains(Country["FRA"]))
```

returns only votes related to France.