Fix inconsistent use of timestamp/date
I previously changed the Xapian index to store the date (without time) in the `timestamp` field, which is a bit confusing. So now there’s `date` (both on the SQL model and in the Xapian index) which is just the date, and there still is `timestamp` which contains the full timestamp (although this isn’t indexed in the Xapian index).