Always include selected options in list of facet options even if there are no matching results
This fixes the issue described in this comment: https://github.com/HowTheyVote/howtheyvote/pull/1224#pullrequestreview-3327530785

For example, if a user selects the filters `country: Russia` and `committee: CONT` and there are no votes that match both filters, we will still include these options in the list because they are selected. This is what most search engines that provide faceted search seem to do in this case, and I think it makes sense.

As discussed before, the ideal solution would be to automatically refresh the available options whenever the selection for any of the other filters changes. This would prevent the situation in the first place, at least in most cases, and I still want to implement that as an additional enhancement on top of this fix.