Handle de-/serialization of facet option values transparently
Previously, a query response would only return raw facet option values as they were indexed (e.g., `FRA` in case of geographic areas). This meant that we had to explicitly deserialize these in the API endpoints in order to return a human-readable label, etc.

This commit changes that by specifying the schema of our index (i.e. the types of the different fields) once. Then we can transparently deserialize values before returning a query response.

This allows us to remove the logic for explicit deserialization from the API endpoint, but the main motivation is to simplify the implementation of a search filters bugfix (in the next commit).