Implement `SearchQuery` class as a container for all search options
There are multiple different places where we need to manipulate URL parameters related to search options such as filters, sorting, pagination, etc. Right now, we pass these options around to the different components (e.g., `SearchFiltersDialog`). These components can then build a new URL with new parameters or without existing parameters. This is error-prone (see for example #1116) and in many cases redundant.

In the future, there will more cases like this:

* Showing the currently active filters at the top of the search results and allowing users to remove filters with a single click
* MEP profiles (which might eventually also get a search/filter feature)

So I've decided to introduce `SearchQuery` as a container class that represents all the search options. This class can be created from a URL, can be easily manipulated, and the converted back to a URL.