Add MEP IDs to search index (to filter votes on MEP profile pages)
This increases index sizes, in case of my local index from ~16 MB to 26 MB. While this is significant as a relative increase, it’s still pretty much irrelevant in absolute numbers.

The other alternative I could think of:

1. Determine the periods during which the MEP was an "active" member (based on group memberships).
2. Filter votes based on these time frames.

However, this would be more complex, especially if we also want to support all the search and filtering features for MEP profiles that we also support for normal searches, so I think this is the better solution.